Refactor DeterminismTest to expose internals as an app.
